audio_to_text_requirements.txt
# 音频转文字工具依赖包
openai-whisper>=20231117
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.24.0
ffmpeg-python>=0.2.0

# 繁简转换（确保输出简体中文）
opencc-python-reimplemented>=0.1.7
